form valuidatiob

+++ main.py
@@ -36,7 +36,7 @@ class DataSet(db.Model):
 class PersonData(db.Model):
   data_set = db.ReferenceProperty(DataSet)
   gender = db.StringProperty(required=True)
-  age = db.FloatProperty(required=True)
+  age = db.IntegerProperty(required=True)
   foot_length = db.FloatProperty(required=True)
   height = db.FloatProperty(required=True)
   stride_length = db.FloatProperty(required=False)
@@ -146,7 +146,7 @@ class FootprintsDataSetHandler(BaseRequestHandler):
   def post(self,key=''):
     data_set = DataSet.get(key);
     gender = self.request.get('gender')
-    age = float(self.request.get('age'))
+    age = int(self.request.get('age'))
     foot_length = float(self.request.get('foot_length'))
     if self.request.get('stride_length'):
       stride_length = float(self.request.get('stride_length'))
